Code refactoring, separated data utils.
