+++ src/view/data/reducer.js
@@ -20,7 +20,7 @@ import {
 	getEditorViewRanges,
 	getEditorViewRoots,
 	getEditorViewTreeDefinition
-} from '../utils';
+} from './utils';
 
 import LocalStorageManager from '../../localstoragemanager';
